examples/backend/python/requirements.txt
flask==3.0.3
pyjwt==2.9.0
eth-account==0.11.3
